authorization implemented

+++ backend/routes/posts.js
@@ -30,7 +30,8 @@ router.post("", checkAuth, multer({storage: storage}).single('image') ,(req, res
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then((createdPost) => {
     res.status(201).json({
@@ -56,14 +57,21 @@ router.put("/:id", checkAuth , multer({storage: storage}).single('image'), (req,
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
 
-  Post.updateOne({ _id: req.params.id }, post).then((result) => {
-    console.log(result);
-    res.status(200).json({
-      message: "Post updated successfully",
-    });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then((result) => {
+    if(result.modifiedCount > 0) {
+      res.status(200).json({
+        message: "Post updated successfully",
+      });
+    } else {
+      res.status(401).json({
+        message: "Not authorized!!",
+      });
+    }
+
   });
 });
 
@@ -108,11 +116,16 @@ router.get("/:id", (req, res) => {
 
 router.delete("/:id", (req, res, next) => {
   // console.log(req.params.id);
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
-    console.log(result);
-    res.status(200).json({
-      message: "Post Deleted Successfully",
-    });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
+    if(result.deletedCount > 0) {
+      res.status(200).json({
+        message: "Post deleted successfully",
+      });
+    } else {
+      res.status(401).json({
+        message: "Not authorized!!",
+      });
+    }
   });
 });
